requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cityscapesscripts==2.2.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
configargparse==1.7
configparse==0.1.5
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.7
cython==3.0.8
debugpy==1.8.1
decorator==5.1.1
deep-sort-realtime==1.3.2
defusedxml==0.7.1
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fastremap==1.14.0
filelock==3.13.1
flask==3.0.0
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.3
gdown==4.7.1
google-auth==2.25.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json-tricks==3.17.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server==2.12.5
jupyter-server-terminals==0.5.2
jupyterlab==4.1.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.3
jupyterlab-widgets==3.0.10
kornia==0.7.0
lazy-loader==0.3
llvmlite==0.41.1
markdown==3.5.1
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmcv-full==1.6.1
model-index==0.1.11
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
nms==0.1.6
notebook==7.1.0
notebook-shim==0.2.4
numba==0.58.1
numpy==1.23.2
nvidia-ml-py==12.535.133
nvitop==1.3.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openxlab==0.0.11
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==4.1.0
prettytable==3.9.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pygments==2.17.2
pyproject==1.3.1
pyquaternion==0.9.9
pysocks==1.7.1
python-json-logger==2.0.7
pytorch-revgrad==0.2.0
pytorch-toolbelt==0.5.2
pytz==2023.3.post1
pywavelets==1.4.1
pyyaml==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
qtpy==2.4.1
qudida==0.0.4
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.18.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
send2trash==1.8.2
setuptools==59.5.0
shapely==2.0.2
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
supervision==0.17.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.1
tomli==2.0.1
torch==1.9.1+cu111
torch-ema==0.3
torchvision==0.10.1+cu111
tornado==6.4
tqdm==4.65.2
traitlets==5.14.1
transformers==4.35.2
types-python-dateutil==2.8.19.20240106
typing==3.7.4.3
typing-extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
werkzeug==3.0.1
widgetsnbextension==4.0.10
yapf==0.40.1
yarl==1.9.4
zipp==3.17.0
